Development

+++ modules/product/product.service.js
@@ -9,6 +9,12 @@ const {
   PRODUCT_NOT_FOUND,
   PRODUCT_HAS_NOT_CHANGED,
 } = require('../../error-messages/products.messages');
+const createTranslations = require('../../utils/createTranslations');
+const {
+  addTranslations,
+  updateTranslations,
+  deleteTranslations,
+} = require('../translations/translations.service');
 const {
   CATEGORY_NOT_FOUND,
 } = require('../../error-messages/category.messages');
@@ -45,8 +51,9 @@ const {
   STATUS_CODES: { FORBIDDEN, NOT_FOUND, BAD_REQUEST },
 } = require('../../consts/status-codes');
 const {
-  HISTORY_ACTIONS: { ADD_PRODUCT, DELETE_PRODUCT, EDIT_PRODUCT },
-} = require('../../consts/history-actions');
+  HISTORY_ACTIONS: { ADD_EVENT, DELETE_EVENT, EDIT_EVENT },
+  HISTORY_NAMES: { PRODUCT_EVENT },
+} = require('../../consts/history-events');
 const {
   generateHistoryObject,
   getChanges,
@@ -122,10 +129,13 @@ class ProductsService {
       .exec();
     const products = await this.getProducts({});
     const sortedByPrices = [...products.items].sort(
-      (a, b) => a.basePrice[1].value - b.basePrice[1].value
+      (a, b) =>
+        a.sizes[a.sizes.length - 1].price[0].value -
+        b.sizes[b.sizes.length - 1].price[0].value
     );
-    const minPrice = sortedByPrices[0].basePrice;
-    const maxPrice = sortedByPrices[sortedByPrices.length - 1].basePrice;
+
+    const minPrice = sortedByPrices[0].sizes[0].price;
+    const maxPrice = sortedByPrices[sortedByPrices.length - 1].sizes[0].price;
 
     return {
       minPrice,
@@ -179,12 +189,16 @@ class ProductsService {
       filter.pattern = { $in: pattern };
     }
     if (price && price.length) {
-      filter.basePrice = {
+      filter.sizes = {
         $elemMatch: {
-          currency: getCurrencySign(currency, UAH, USD),
-          value: {
-            $gte: price[0],
-            $lte: price[1],
+          price: {
+            $elemMatch: {
+              currency: getCurrencySign(currency, UAH, USD),
+              value: {
+                $gte: price[0],
+                $lte: price[1],
+              },
+            },
           },
         },
       };
@@ -192,8 +206,11 @@ class ProductsService {
     return filter;
   }
 
-  async getProducts({ filter, skip, limit, sort, search }) {
+  async getProducts({ filter, skip, limit, sort={rate: -1}, search }) {
     const filters = this.filterItems(filter);
+    const sortValue = (Object.keys(sort)).includes('basePrice') ? {
+      'sizes.price.value': sort.basePrice
+    } : sort;
     if (!(!search || search.trim().length === 0)) {
       filters.$or = [
         {
@@ -208,10 +225,11 @@ class ProductsService {
         },
       ];
     }
+
     const items = await Product.find(filters)
+      .sort(sortValue)
       .skip(skip)
       .limit(limit)
-      .sort(sort)
       .exec();
 
     const count = await Product.find(filters)
@@ -231,6 +249,9 @@ class ProductsService {
     primary,
     { _id: adminId }
   ) {
+    const matchPrimaryInUpload = filesToUpload.filter(
+      item => item.large === productData.images[0].primary.large
+    );
     productData.images = {
       primary: {
         large: LARGE_SAD_BACKPACK,
@@ -239,18 +260,7 @@ class ProductsService {
         thumbnail: THUMBNAIL_SAD_BACKPACK,
       },
     };
-    if (filesToUpload.length) {
-      productData.images.additional = [];
-    } else {
-      productData.images.additional = [
-        {
-          large: LARGE_SAD_BACKPACK,
-          medium: MEDIUM_SAD_BACKPACK,
-          small: SMALL_SAD_BACKPACK,
-          thumbnail: THUMBNAIL_SAD_BACKPACK,
-        },
-      ];
-    }
+
     const product = await Product.findById(id)
       .lean()
       .exec();
@@ -264,17 +274,19 @@ class ProductsService {
       if (primary?.large) {
         productData.images.primary = primary;
       } else {
-        await uploadService.deleteFiles(
-          Object.values(product.images.primary).filter(
-            item => typeof item === 'string'
-          )
-        );
-        const uploadResult = await uploadService.uploadFiles(primary);
+        if (!matchPrimaryInUpload.length)
+          await uploadService.deleteFiles(
+            Object.values(product.images.primary).filter(
+              item => typeof item === 'string'
+            )
+          );
+        const uploadResult = await uploadService.uploadFiles([primary]);
         const imagesResults = await uploadResult[0];
         productData.images.primary = imagesResults?.fileNames;
       }
     }
     if (filesToUpload.length) {
+      productData.images.additional = [];
       const previousImagesLinks = [];
       const newFiles = [];
       filesToUpload.forEach(e => {
@@ -288,24 +300,39 @@ class ProductsService {
       const imagesResults = await Promise.allSettled(newUploadResult);
       const additional = imagesResults.map(res => res?.value?.fileNames);
       productData.images.additional = [...additional, ...previousImagesLinks];
+    } else {
+      productData.images.additional = [
+        {
+          large: LARGE_SAD_BACKPACK,
+          medium: MEDIUM_SAD_BACKPACK,
+          small: SMALL_SAD_BACKPACK,
+          thumbnail: THUMBNAIL_SAD_BACKPACK,
+        },
+      ];
     }
     const { basePrice } = productData;
     productData.basePrice = await calculateBasePrice(basePrice);
     productData.sizes = await finalPriceCalculation(productData);
-    if (productData) {
-      const { beforeChanges, afterChanges } = getChanges(product, productData);
 
-      const historyRecord = generateHistoryObject(
-        EDIT_PRODUCT,
-        product.model,
-        product.name[UA].value,
-        product._id,
-        beforeChanges,
-        afterChanges,
-        adminId
-      );
-      await addHistoryRecord(historyRecord);
-    }
+    const { beforeChanges, afterChanges } = getChanges(product, productData);
+    const historyEvent = {
+      action: EDIT_EVENT,
+      historyName: PRODUCT_EVENT,
+    };
+    const historyRecord = generateHistoryObject(
+      historyEvent,
+      product.model,
+      product.name[UA].value,
+      product._id,
+      beforeChanges,
+      afterChanges,
+      adminId
+    );
+    await addHistoryRecord(historyRecord);
+    await updateTranslations(
+      product.translationsKey,
+      createTranslations(productData)
+    );
 
     return Product.findByIdAndUpdate(id, productData, {
       new: true,
@@ -329,12 +356,19 @@ class ProductsService {
     };
 
     productData.sizes = await finalPriceCalculation(productData);
+    productData.translationsKey = await addTranslations(
+      createTranslations(productData)
+    );
 
     const newProduct = await new Product(productData).save();
 
     if (productData) {
+      const historyEvent = {
+        action: ADD_EVENT,
+        historyName: PRODUCT_EVENT,
+      };
       const historyRecord = generateHistoryObject(
-        ADD_PRODUCT,
+        historyEvent,
         newProduct.model?._id,
         newProduct.name[UA].value,
         newProduct._id,
@@ -369,20 +403,30 @@ class ProductsService {
     const product = await Product.findById(id)
       .lean()
       .exec();
+
     if (!product) {
       throw new RuleError(PRODUCT_NOT_FOUND, NOT_FOUND);
     }
+
     const { images } = product;
     const { primary, additional } = images;
-    const additionalImagesToDelete = Object.assign(...additional);
+    const additionalImagesToDelete =
+      typeof additional[0] === 'object'
+        ? additional.map(img => [...Object.values(img)]).flat()
+        : [];
+
     const deletedImages = await uploadService.deleteFiles([
       ...Object.values(primary),
-      ...Object.values(additionalImagesToDelete),
+      ...additionalImagesToDelete,
     ]);
 
     if (await Promise.allSettled(deletedImages)) {
+      const historyEvent = {
+        action: DELETE_EVENT,
+        historyName: PRODUCT_EVENT,
+      };
       const historyRecord = generateHistoryObject(
-        DELETE_PRODUCT,
+        historyEvent,
         product.model,
         product.name[UA].value,
         product._id,
@@ -406,6 +450,8 @@ class ProductsService {
         [],
         adminId
       );
+
+      await deleteTranslations(product.translationsKey);
 
       await addHistoryRecord(historyRecord);
